requirements.txt
###### Requirements without Version Specifiers ######

django
Pillow
django_rest_framework
django-allauth